refactor(tabs): don't use barrels
