+++ src/lib/src/index.ts
@@ -2,6 +2,6 @@
  * Public API Surface of ngx-foundation-sites
  */
 
-export { FasTabComponent } from './lib/tabs/tab';
-export { FasTabsComponent } from './lib/tabs/tabs';
-export { FasTabsModule } from './lib/tabs';
+export { FasTabComponent } from './lib/tabs/tab/tab.component';
+export { FasTabsComponent } from './lib/tabs/tabs/tabs.component';
+export { FasTabsModule } from './lib/tabs/tabs.module';
