--- src/lib/src/lib/tabs/index.ts
@@ -1,1 +1,0 @@
-export { FasTabsModule } from './tabs.module';
