--- src/lib/src/lib/tabs/tab/index.ts
@@ -1,2 +1,0 @@
-export * from './tab.component';
-export * from './tab.module';
