--- src/lib/src/lib/tabs/tabs/index.ts
@@ -1,2 +1,0 @@
-export * from './tabs.component';
-export * from './tabs.module';
